Creation of "lattice" and "physics" packages
