+++ pyat/at/lattice/__init__.py
@@ -1,0 +1,8 @@
+"""
+Helper functions for working with AT lattices.
+
+A lattice as understood by pyAT is any sequence of elements.  These functions
+are useful for working with these sequences.
+"""
+
+from .utils import *
